# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Plan Management API
  version: 1.0.0
  description: 주간 계획표 관리 서비스의 API 문서입니다.
servers:
  - url: http://localhost:4000/api
    description: Local server
paths:
...
  /folders:
    get:
      summary: "폴더 목록 조회"
      tags: [Folders]
      responses:
        "200":
          description: "폴더 목록 조회 성공"
        "500":
          description: "서버 오류"
    post:
      summary: "폴더 추가"
      tags: [Folders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                folder_name:
                  type: string
                  example: "업무폴더"
      responses:
        "201":
          description: "폴더 추가 성공"
        "400":
          description: "잘못된 요청"
        "500":
          description: "서버 오류"
  /folders/{fid}:
    put:
      summary: "폴더 이름 수정"
      tags: [Folders]
      parameters:
        - in: path
          name: fid
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                folder_name:
                  type: string
                  example: "새 폴더명"
      responses:
        "200":
          description: "폴더 수정 완료"
        "500":
          description: "서버 오류"
    delete:
      summary: "폴더 삭제"
      tags: [Folders]
      parameters:
        - in: path
          name: fid
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: "폴더 삭제 완료"
        "404":
          description: "폴더를 찾을 수 없음"
        "500":
          description: "서버 오류"
  /folders/{fid}/plans:
    get:
      summary: "계획표 목록 조회"
      tags: [Plans]
      parameters:
        - in: path
          name: fid
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "계획표 목록 조회 성공"
        "500":
          description: "서버 오류"
    post:
      summary: "계획표 추가"
      tags: [Plans]
      parameters:
        - in: path
          name: fid
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                week:
                  type: string
                table:
                  type: array
                  items:
                    type: array
                    items:
                      type: string
      responses:
        "201":
          description: "계획표 추가 완료"
        "500":
          description: "서버 오류"
  /folders/{fid}/plans/{pid}:
    put:
      summary: "계획표 수정"
      tags: [Plans]
      parameters:
        - in: path
          name: fid
          required: true
          schema:
            type: integer
        - in: path
          name: pid
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                week:
                  type: string
                table:
                  type: array
                  items:
                    type: array
                    items:
                      type: string
      responses:
        "200":
          description: "계획표 수정 완료"
        "404":
          description: "해당 계획표 없음"
        "500":
          description: "서버 오류"
    delete:
      summary: "계획표 삭제"
      tags: [Plans]
      parameters:
        - in: path
          name: fid
          required: true
          schema:
            type: integer
        - in: path
          name: pid
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "계획표 삭제 완료"
        "404":
          description: "계획표를 찾을 수 없음"
        "500":
          description: "서버 오류"
  /folders/{fid}/plans/{pid}/projects:
    get:
      summary: "계획표의 프로젝트 및 유저 정보 조회"
      tags: [Plans]
      parameters:
        - in: path
          name: fid
          required: true
          schema:
            type: integer
        - in: path
          name: pid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "프로젝트 및 유저 정보 조회 성공"
        "500":
          description: "서버 오류"
  /folders/{fid}/plans/{pid}/feedback:
    put:
      summary: "피드백 수정"
      tags: [Plans]
      parameters:
        - in: path
          name: fid
          required: true
          schema:
            type: integer
        - in: path
          name: pid
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                feedback:
                  type: string
      responses:
        "200":
          description: "Feedback 수정 완료"
        "500":
          description: "서버 오류"
  /search:
    get:
      summary: "키워드로 사용자 데이터 검색"
      description: |
        현재 사용자가 소유한 폴더, 계획표, 프로젝트를 키워드로 검색합니다.
        프로젝트는 project_name, last_week, this_week 필드도 검색에 포함됩니다.
      tags: [Search]
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
          required: true
          description: "검색어"
      responses:
        "200":
          description: "검색 결과 반환 성공"
          content:
            application/json:
              schema:
                type: object
                properties:
                  folders:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        folder_name:
                          type: string
                  plans:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        title:
                          type: string
                        fid:
                          type: integer
                  projects:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        project_name:
                          type: string
                        last_week:
                          type: string
                        this_week:
                          type: string
                        pid:
                          type: integer
                        fid:
                          type: integer
        "400":
          description: "검색어 누락"
        "500":
          description: "서버 오류"
...